# church-app/public/openapi.yaml
openapi: 3.0.3
info:
  title: Church App — API
  version: "1.0.0"
  description: >
    Documentation des points d'accès backend de l'application Church App
    (Centre International des Vainqueurs).


    Le backend de ce projet est **Supabase** : il n'y a pas de serveur API
    "maison" pour les données — le frontend appelle directement l'API REST
    auto-générée par Supabase (PostgREST) au-dessus des tables Postgres,
    protégées par des policies RLS. Cette spec documente ces tables comme des
    endpoints, plus la (les) route(s) serveur Next.js propres au projet.


    **Pour tester une table Supabase dans Swagger UI :**
    1. Cliquez sur "Authorize".
    2. Renseignez `apikey` avec la clé anon (`NEXT_PUBLIC_SUPABASE_ANON_KEY`
       dans `.env.local`).
    3. Renseignez `bearerAuth` avec un access token JWT d'un utilisateur
       connecté (récupérable via `supabase.auth.getSession()` dans la console
       du navigateur sur l'app, ou via Supabase Dashboard > Authentication).
    4. Les policies RLS s'appliquent normalement : une requête échouera ou ne
       renverra rien si le rôle du token ne le permet pas.


    ⚠️ Doc vivante — à mettre à jour à chaque table/colonne/policy ajoutée ou
    modifiée. Voir aussi [docs/BACKEND.md](./docs/BACKEND.md) pour le détail
    des policies RLS, fonctions SQL et historique des correctifs.
servers:
  - url: https://nmqctnkfqzqhdgaxneet.supabase.co/rest/v1
    description: Supabase REST (PostgREST) — tables de l'application
...
security:
  - apiKeyAuth: []
    bearerAuth: []
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: apikey
      description: Clé anon Supabase (NEXT_PUBLIC_SUPABASE_ANON_KEY)
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token de l'utilisateur connecté (auth.uid())
...
  schemas:
...
    Erreur:
      type: object
      properties:
        error: { type: string }
paths:
...
  /api/admin/utilisateurs/{id}:
    delete:
      tags: [admin]
      summary: Supprimer définitivement un compte (auth + profil)
      description: >
        Route serveur Next.js (PAS Supabase direct) — utilise la clé
        `service_role` côté serveur pour contourner RLS et appeler
        `auth.admin.deleteUser()`. Vérifie elle-même que l'appelant est
        authentifié et admin/pasteur. Voir
        [src/app/api/admin/utilisateurs/[id]/route.ts](../src/app/api/admin/utilisateurs/[id]/route.ts).
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Compte supprimé
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
        "400":
          description: Tentative d'auto-suppression
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Erreur' }
        "401":
          description: Non authentifié
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Erreur' }
        "403":
          description: Réservé aux administrateurs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Erreur' }